# specs/openapi.yaml
openapi: 3.0.3
info:
  title: FaMoney API
  description: 공동 지출 원장 웹 애플리케이션 API
  version: 1.0.0
  contact:
    name: FaMoney Team
servers:
  - url: http://localhost:8080
    description: Development server
  - url: https://api.famoney.app
    description: Production server
...
paths:
...
  /api/v1/ledgers/{ledgerId}/expenses:
    post:
      tags: [Expenses]
      summary: 지출 기록 생성
      operationId: createExpense
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/LedgerIdParam'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateExpenseRequest'
      responses:
        '201':
          description: 생성 성공
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ExpenseResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'

    get:
      tags: [Expenses]
      summary: 지출 기록 목록 조회
      operationId: getExpenses
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/LedgerIdParam'
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/SizeParam'
        - $ref: '#/components/parameters/SortParam'
        - name: startDate
          in: query
          schema:
            type: string
            format: date
          description: 조회 시작일
        - name: endDate
          in: query
          schema:
            type: string
            format: date
          description: 조회 종료일
        - name: category
          in: query
          schema:
            type: string
          description: 카테고리 필터
        - name: createdBy
          in: query
          schema:
            type: string
            format: uuid
          description: 작성자 필터
      responses:
        '200':
          description: 성공
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ExpenseListResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /api/v1/ledgers/{ledgerId}/expenses/{expenseId}:
    get:
      tags: [Expenses]
      summary: 지출 기록 상세 조회
      operationId: getExpense
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/LedgerIdParam'
        - $ref: '#/components/parameters/ExpenseIdParam'
      responses:
        '200':
          description: 성공
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ExpenseResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'

    patch:
      tags: [Expenses]
      summary: 지출 기록 수정
      operationId: updateExpense
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/LedgerIdParam'
        - $ref: '#/components/parameters/ExpenseIdParam'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateExpenseRequest'
      responses:
        '200':
          description: 성공
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ExpenseResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'

    delete:
      tags: [Expenses]
      summary: 지출 기록 삭제
      operationId: deleteExpense
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/LedgerIdParam'
        - $ref: '#/components/parameters/ExpenseIdParam'
      responses:
        '204':
          description: 삭제 성공
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'

  # ============================================
  # Categories
  # ============================================
  /api/v1/ledgers/{ledgerId}/categories:
    get:
      tags: [Expenses]
      summary: 카테고리 목록 조회
      operationId: getCategories
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/LedgerIdParam'
      responses:
        '200':
          description: 성공
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CategoryListResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'

    post:
      tags: [Expenses]
      summary: 커스텀 카테고리 생성
      operationId: createCategory
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/LedgerIdParam'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateCategoryRequest'
      responses:
        '201':
          description: 생성 성공
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CategoryResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
components:
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  # ============================================
  # Parameters
  # ============================================
  parameters:
    LedgerIdParam:
      name: ledgerId
      in: path
      required: true
      schema:
        type: string
        format: uuid
      description: 원장 ID
...
    ExpenseIdParam:
      name: expenseId
      in: path
      required: true
      schema:
        type: string
        format: uuid
      description: 지출 ID
    PageParam:
      name: page
      in: query
      schema:
        type: integer
        default: 0
        minimum: 0
      description: 페이지 번호 (0부터 시작)
    SizeParam:
      name: size
      in: query
      schema:
        type: integer
        default: 20
        minimum: 1
        maximum: 100
      description: 페이지 크기
    SortParam:
      name: sort
      in: query
      schema:
        type: string
        default: expenseDate,desc
      description: "정렬 (예: expenseDate,desc)"

  # ============================================
  # Responses
  # ============================================
  responses:
    BadRequest:
      description: 잘못된 요청
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Unauthorized:
      description: 인증 실패
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Forbidden:
      description: 권한 없음
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    NotFound:
      description: 리소스를 찾을 수 없음
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'

  # ============================================
  # Schemas
  # ============================================
  schemas:
...
    ErrorResponse:
      type: object
      required:
        - timestamp
        - status
        - error
        - message
        - path
      properties:
        timestamp:
          type: string
          format: date-time
        status:
          type: integer
        error:
          type: string
        message:
          type: string
        path:
          type: string
        details:
          type: array
          items:
            $ref: '#/components/schemas/FieldError'
    FieldError:
      type: object
      properties:
        field:
          type: string
        message:
          type: string
    PageInfo:
      type: object
      properties:
        page:
          type: integer
        size:
          type: integer
        totalElements:
          type: integer
          format: int64
        totalPages:
          type: integer

    # ------------------------------------------
    # Auth
    # ------------------------------------------
...
    UserSummary:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        profileImage:
          type: string
          nullable: true

    # ------------------------------------------
    # Expense
    # ------------------------------------------
    CreateExpenseRequest:
      type: object
      required:
        - amount
        - description
        - expenseDate
      properties:
        amount:
          type: number
          format: double
          minimum: 0
          exclusiveMinimum: true
        description:
          type: string
          minLength: 1
          maxLength: 255
        category:
          type: string
          maxLength: 50
        expenseDate:
          type: string
          format: date
        paymentMethod:
          type: string
          maxLength: 50
    UpdateExpenseRequest:
      type: object
      properties:
        amount:
          type: number
          format: double
          minimum: 0
          exclusiveMinimum: true
        description:
          type: string
          minLength: 1
          maxLength: 255
        category:
          type: string
          maxLength: 50
        expenseDate:
          type: string
          format: date
        paymentMethod:
          type: string
          maxLength: 50
    ExpenseResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        amount:
          type: number
          format: double
        description:
          type: string
        category:
          type: string
          nullable: true
        expenseDate:
          type: string
          format: date
        paymentMethod:
          type: string
          nullable: true
        createdBy:
          $ref: '#/components/schemas/UserSummary'
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
    ExpenseListResponse:
      type: object
      properties:
        content:
          type: array
          items:
            $ref: '#/components/schemas/ExpenseResponse'
        page:
          $ref: '#/components/schemas/PageInfo'
        summary:
          $ref: '#/components/schemas/ExpenseListSummary'
    ExpenseListSummary:
      type: object
      properties:
        totalAmount:
          type: number
          format: double
        count:
          type: integer

    # ------------------------------------------
    # Category
    # ------------------------------------------
    CategoryResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        color:
          type: string
          description: "HEX 색상 코드 (예: #FF5733)"
        isDefault:
          type: boolean
    CategoryListResponse:
      type: object
      properties:
        content:
          type: array
          items:
            $ref: '#/components/schemas/CategoryResponse'
    CreateCategoryRequest:
      type: object
      required:
        - name
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 50
        color:
          type: string
          pattern: '^#[0-9A-Fa-f]{6}$'